@@ -1,6 +1,6 @@
 {
   "name": "@opentelemetry/instrumentation-user-interaction",
-  "version": "0.38.0",
+  "version": "0.39.0",
   "description": "OpenTelemetry instrumentation for user interactions as click events in a web application",
   "main": "build/src/index.js",
   "module": "build/esm/index.js",
@@ -14,10 +14,10 @@
     "precompile": "tsc --version && lerna run version:update --scope @opentelemetry/instrumentation-user-interaction --include-dependencies",
     "prewatch": "npm run precompile",
     "version:update": "node ../../../scripts/version-update.js",
-    "compile": "npm run version:update && tsc --build tsconfig.json tsconfig.esm.json tsconfig.esnext.json",
+    "compile": "tsc --build tsconfig.json tsconfig.esm.json tsconfig.esnext.json",
     "prepublishOnly": "npm run compile",
     "tdd": "karma start",
-    "test:browser": "nyc karma start --single-run",
+    "test:browser": "karma start --single-run",
     "watch": "tsc --build --watch tsconfig.json tsconfig.esm.json tsconfig.esnext.json"
   },
   "keywords": [
@@ -48,40 +48,42 @@
     "access": "public"
   },
   "devDependencies": {
-    "@babel/core": "7.22.17",
-    "@jsdevtools/coverage-istanbul-loader": "3.0.5",
+    "@babel/core": "7.24.6",
+    "@babel/preset-env": "7.24.6",
     "@opentelemetry/api": "^1.3.0",
     "@opentelemetry/context-zone-peer-dep": "^1.8.0",
-    "@opentelemetry/instrumentation-xml-http-request": "^0.51.0",
+    "@opentelemetry/instrumentation-xml-http-request": "^0.52.0",
     "@opentelemetry/sdk-trace-base": "^1.8.0",
-    "@types/jquery": "3.5.20",
-    "@types/mocha": "7.0.2",
+    "@types/jquery": "3.5.30",
+    "@types/mocha": "10.0.6",
     "@types/node": "18.6.5",
-    "@types/sinon": "10.0.18",
-    "@types/webpack-env": "1.16.2",
-    "babel-loader": "8.2.2",
-    "karma": "6.3.16",
-    "karma-chrome-launcher": "^3.1.1",
-    "karma-coverage-istanbul-reporter": "3.0.3",
+    "@types/sinon": "17.0.3",
+    "@types/webpack-env": "1.16.3",
+    "assert": "2.0.0",
+    "babel-loader": "8.3.0",
+    "babel-plugin-istanbul": "6.1.1",
+    "cross-var": "1.1.0",
+    "karma": "6.4.3",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage": "2.2.1",
     "karma-jquery": "0.2.4",
     "karma-mocha": "2.0.1",
     "karma-spec-reporter": "0.0.36",
-    "karma-webpack": "4.0.2",
-    "mocha": "7.2.0",
+    "karma-webpack": "5.0.1",
+    "mocha": "10.2.0",
     "nyc": "15.1.0",
-    "rimraf": "5.0.5",
-    "sinon": "15.2.0",
-    "ts-loader": "8.3.0",
+    "sinon": "15.1.2",
+    "ts-loader": "9.5.1",
     "ts-mocha": "10.0.0",
     "typescript": "4.4.4",
-    "webpack": "4.46.0",
-    "webpack-cli": "4.7.2",
-    "webpack-merge": "5.8.0",
+    "webpack": "5.89.0",
+    "webpack-cli": "5.1.4",
+    "webpack-merge": "5.10.0",
     "zone.js": "^0.11.4 || ^0.13.0 || ^0.14.0"
   },
   "dependencies": {
     "@opentelemetry/core": "^1.8.0",
-    "@opentelemetry/instrumentation": "^0.51.0",
+    "@opentelemetry/instrumentation": "^0.52.0",
     "@opentelemetry/sdk-trace-web": "^1.8.0"
   },
   "peerDependencies": {
