@@ -13,5 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './GrpcCensusPropagator';
+declare module 'systeminformation/lib/network' {
+  export { networkStats } from 'systeminformation';
+}
