@@ -14,17 +14,17 @@
  * limitations under the License.
  */
 
-import * as api from '@opentelemetry/api';
-import * as metrics from '@opentelemetry/sdk-metrics';
+import { Meter, diag, metrics } from '@opentelemetry/api';
+import { MeterProvider } from '@opentelemetry/sdk-metrics';
 
-import { VERSION } from './version';
+import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
 
 /**
  * Metrics Collector Configuration
  */
 export interface MetricsCollectorConfig {
   // Meter Provider
-  meterProvider?: metrics.MeterProvider;
+  meterProvider?: MeterProvider;
   // Character to be used to join metrics - default is "."
   metricNameSeparator?: string;
   // Name of component
@@ -33,24 +33,23 @@ export interface MetricsCollectorConfig {
   url?: string;
 }
 
-const DEFAULT_NAME = '@opentelemetry/host-metrics';
+const DEFAULT_NAME = PACKAGE_NAME;
 
 /**
  * Base Class for metrics
  */
 export abstract class BaseMetrics {
-  protected _logger = api.diag;
-  protected _meter: api.Meter;
+  protected _logger = diag;
+  protected _meter: Meter;
   private _name: string;
 
   constructor(config: MetricsCollectorConfig) {
     this._name = config.name || DEFAULT_NAME;
-    const meterProvider =
-      config.meterProvider || api.metrics.getMeterProvider();
+    const meterProvider = config.meterProvider || metrics.getMeterProvider();
     if (!config.meterProvider) {
       this._logger.warn('No meter provider, using default');
     }
-    this._meter = meterProvider.getMeter(this._name, VERSION);
+    this._meter = meterProvider.getMeter(this._name, PACKAGE_VERSION);
   }
 
   /**
