@@ -16,7 +16,7 @@
 
 import { Logger, logs } from '@opentelemetry/api-logs';
 import TransportStream = require('winston-transport');
-import { VERSION } from './version';
+import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
 import { emitLogRecord } from './utils';
 
 export class OpenTelemetryTransportV3 extends TransportStream {
@@ -24,7 +24,7 @@ export class OpenTelemetryTransportV3 extends TransportStream {
 
   constructor(options?: TransportStream.TransportStreamOptions) {
     super(options);
-    this._logger = logs.getLogger('@opentelemetry/winston-transport', VERSION);
+    this._logger = logs.getLogger(PACKAGE_NAME, PACKAGE_VERSION);
   }
 
   public override log(info: any, callback: () => void) {
