@@ -17,7 +17,7 @@ import { strict as assert } from 'assert';
 import * as http from 'http';
 import { AddressInfo } from 'net';
 
-import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
+import { SEMATTRS_MESSAGING_SYSTEM } from '@opentelemetry/semantic-conventions';
 import { ReadableSpan } from '@opentelemetry/sdk-trace-base';
 import { getTestSpans } from '@opentelemetry/contrib-test-utils';
 
@@ -57,7 +57,7 @@ export const createServerInstance = (server?: http.Server) => {
 
 export const getSocketIoSpans = (): ReadableSpan[] =>
   getTestSpans().filter(
-    s => s.attributes[SemanticAttributes.MESSAGING_SYSTEM] === 'socket.io'
+    s => s.attributes[SEMATTRS_MESSAGING_SYSTEM] === 'socket.io'
   ) as ReadableSpan[];
 
 export const expectSpan = (
