@@ -27,7 +27,7 @@ import {
 import { AddressFamily } from './enums/AddressFamily';
 import { DnsInstrumentationConfig } from './types';
 import * as utils from './utils';
-import { VERSION } from './version';
+import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
 import {
   LookupCallbackSignature,
   LookupPromiseSignature,
@@ -40,7 +40,7 @@ export class DnsInstrumentation extends InstrumentationBase {
   protected override _config!: DnsInstrumentationConfig;
 
   constructor(config: DnsInstrumentationConfig = {}) {
-    super('@opentelemetry/instrumentation-dns', VERSION, config);
+    super(PACKAGE_NAME, PACKAGE_VERSION, config);
   }
 
   init(): (
