@@ -33,7 +33,7 @@ import {
 import { safeExecuteInTheMiddle } from '@opentelemetry/instrumentation';
 import { endSpan } from './utils';
 import { defaultDbStatementSerializer } from '@opentelemetry/redis-common';
-import { VERSION } from './version';
+import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
 
 const DEFAULT_CONFIG: IORedisInstrumentationConfig = {
   requireParentSpan: true,
@@ -44,8 +44,8 @@ export class IORedisInstrumentation extends InstrumentationBase {
 
   constructor(config: IORedisInstrumentationConfig = {}) {
     super(
-      '@opentelemetry/instrumentation-ioredis',
-      VERSION,
+      PACKAGE_NAME,
+      PACKAGE_VERSION,
       Object.assign({}, DEFAULT_CONFIG, config)
     );
   }
