@@ -40,7 +40,8 @@ const content = `/*
  */
 
 // this is autogenerated file, see scripts/version-update.js
-export const VERSION = '${pjson.version}';
+export const PACKAGE_VERSION = '${pjson.version}';
+export const PACKAGE_NAME = '${pjson.name}';
 `;
 
 const fileUrl = path.join(appRoot, 'src', 'version.ts')
